refactor: transform some styles to component
